Moved columns logic into render_json and render_table.
